Code cleanup, interface changes

+++ src/jvmMain/java/game/output/ui/IElement.java
@@ -11,9 +11,9 @@ public abstract class IElement {
     public int x;
     public int y;
 
-    public IElement(double x, double y) {
-        this.x = (int) x;
-        this.y = (int) y;
+    public IElement(int x, int y) {
+        this.x = x;
+        this.y = y;
     }
 
     public abstract boolean press(double x, double y);
